Split out source files and move integration tests
Integration tests should be in a separate subdirectory to test that the API was
properly exported.
